SDK-1472: Move profile sandbox client into Yoti\Sandbox\Profile namespace
