+++ .gitignore
@@ -14,3 +14,5 @@ coverage
 .scannerwork
 
 .DS_Store
+cghooks.lock
+
